+++ packages/blade/src/components/Input/PasswordInput/__tests__/PasswordInput.native.test.tsx
@@ -63,7 +63,6 @@ describe('<PasswordInput />', () => {
     const input = getByPlaceholderText(placeholder);
 
     // we assume auto focus is working with this prop in place, no simple way of asserting on focus otherwise
-    // @ts-expect-error TS typings not being picked from library
     expect(input).toHaveProp('autoFocus', true);
   });
 
@@ -165,13 +164,11 @@ describe('<PasswordInput />', () => {
     const input = getByPlaceholderText(placeholder);
     const button = getByRole('button');
 
-    // @ts-expect-error TS typings not being picked from library
     expect(input).toHaveProp('secureTextEntry', true);
     getByLabelText('Show password');
 
     fireEvent.press(button);
 
-    // @ts-expect-error TS typings not being picked from library
     expect(input).toHaveProp('secureTextEntry', false);
     getByLabelText('Hide password');
   });