+++ packages/blade/src/components/Input/TextArea/__tests__/TextArea.native.test.tsx
@@ -67,7 +67,6 @@ describe('<TextArea />', () => {
     const input = getByPlaceholderText(placeholder);
 
     // we assume auto focus is working with this prop in place, no simple way of asserting on focus otherwise
-    // @ts-expect-error TS typings not being picked from library
     expect(input).toHaveProp('autoFocus', true);
   });