+++ packages/blade/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @razorpay/blade
 
+## 5.2.1
+
+### Patch Changes
+
+- 9966931: chore: fix dom nesting in form label component
+- e660831: fix: change acceptable BaseInput `type` from `numeric` to `number`
+
+## 5.2.0
+
+### Minor Changes
+
+- d03de10: feat(Alert): update `isFullWidth` to make inline borderless alerts on desktop
+
+  > **Warning**
+  >
+  > `isBorderless` prop is removed and its usage is now replaced by `isFullWidth`. The layout is updated to match the designs and is now centered on desktop resolutions.
+
+  ### Steps for migration:
+
+  ```diff
+  <Alert
+  - isBorderless
+  + isFullWidth
+  />
+  ```
+
 ## 5.1.5
 
 ### Patch Changes